Extract DynamoDBTableTestBase from DynamoDBRecordBuilderTest
