+++ java/dynamodb-tools/src/test/java/sleeper/dynamodb/tools/DynamoDBTableTestBase.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright 2022-2023 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package sleeper.dynamodb.tools;
+
+import com.amazonaws.services.dynamodbv2.model.AttributeDefinition;
+import com.amazonaws.services.dynamodbv2.model.KeySchemaElement;
+import com.amazonaws.services.dynamodbv2.model.KeyType;
+import com.amazonaws.services.dynamodbv2.model.ScalarAttributeType;
+import org.junit.jupiter.api.AfterEach;
+
+import java.util.Arrays;
+
+import static sleeper.dynamodb.tools.DynamoDBUtils.initialiseTable;
+
+public class DynamoDBTableTestBase extends DynamoDBTestBase {
+    public static final String TEST_KEY = "test-key";
+    public static final String TEST_VALUE = "test-value";
+    public static final String TEST_TABLE_NAME = "dynamodb-tools-test-table";
+
+    @AfterEach
+    public void tearDown() {
+        dynamoDBClient.deleteTable(TEST_TABLE_NAME);
+    }
+
+    public static void createStringTable() {
+        createTable(ScalarAttributeType.S);
+    }
+
+    public static void createNumericTable() {
+        createTable(ScalarAttributeType.N);
+    }
+
+    public static void createTable(ScalarAttributeType valueType) {
+        initialiseTable(dynamoDBClient, TEST_TABLE_NAME,
+                Arrays.asList(
+                        new AttributeDefinition(TEST_KEY, ScalarAttributeType.S),
+                        new AttributeDefinition(TEST_VALUE, valueType)),
+                Arrays.asList(
+                        new KeySchemaElement(TEST_KEY, KeyType.HASH),
+                        new KeySchemaElement(TEST_VALUE, KeyType.RANGE)));
+    }
+}
